pyproject.toml
[project]
name = "wackelberry-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Marc Osterland",email = "marc.osterland@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)"
]

[tool.poetry]
packages = [{include = "wackelberry_bot", from = "src"}]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
